requirements.txt
agate==1.9.1
annotated-types==0.7.0
attrs==24.2.0
babel==2.16.0
beautifulsoup4==4.12.2
blinker==1.4
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
command-not-found==0.3
cryptography==3.4.8
daff==1.3.46
dbt-adapters==1.7.2
dbt-common==1.11.0
dbt-core==1.9.0b2
dbt-extractor==0.5.1
dbt-postgres==1.8.2
dbt-semantic-interfaces==0.7.4
dbus-python==1.2.18
deepdiff==7.0.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
docutils==0.21.2
greenlet==3.0.3
httplib2==0.20.2
idna==3.10
importlib-metadata==6.11.0
isodate==0.6.1
jeepney==0.7.1
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
leather==0.4.0
lockfile==0.12.2
luigi==3.5.2
MarkupSafe==3.0.2
mashumaro==3.14
more-itertools==8.10.0
msgpack==1.1.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.1
oauthlib==3.2.0
ordered-set==4.1.0
packaging==24.1
pandas==2.2.2
parsedatetime==2.6
pathspec==0.12.1
playwright==1.46.0
protobuf==4.25.5
psycopg2==2.9.9
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pyee==11.1.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu2
python-daemon==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
SecretStorage==3.3.1
sentry-sdk==2.17.0
six==1.16.0
snowplow-tracker==1.0.3
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.1
systemd-python==234
tenacity==8.5.0
text-unidecode==1.3
tornado==6.4.1
types-requests==2.32.0.20241016
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.2.3
wadllib==1.3.6
zipp==1.0.0
